Develop

+++ src/Pages/ManagementProduct.jsx
@@ -1,15 +1,18 @@
-import { Box, Button, Center, Heading, Icon, Input, InputGroup, InputRightElement, NumberInput, NumberInputField, NumberInputStepper, Select, Table, TableContainer, Tbody, Text, Th, Thead, Tr } from '@chakra-ui/react'
+import { Box, Button, Center, Heading, Icon, Input, InputGroup, InputRightElement, NumberInput, NumberInputField, NumberInputStepper, Select, Table, TableContainer, Tbody, Text, Th, Thead, Tr, Td } from '@chakra-ui/react'
 import React, { useEffect, useState } from 'react'
 import MenuManagement from '../Components/MenuManagement'
 import { AiOutlinePlusSquare, AiOutlineEdit, AiOutlineDelete } from 'react-icons/ai'
 import ModalAddProduct from '../Components/ModalAddProduct'
 import { useSelector, useDispatch } from 'react-redux'
 import ModalEditProduct from '../Components/ModalEditProduct'
-import { deleteProductAction } from '../redux/actions'
+import { deleteProductAction, getProductWarehouseAction } from '../redux/actions'
 import { Pagination } from '@mantine/core'
 import axios from 'axios'
 import { API_URL } from '../helper'
 import LoadingPage from './LoadingPage'
+import GoOnTop from '../Components/GoOnTop'
+import BtnOnTop from '../Components/BtnOnTop'
+import Swal from 'sweetalert2'
 
 const ManagementProduct = (props) => {
 
@@ -21,45 +24,41 @@ const ManagementProduct = (props) => {
     const [dataProduct, setDataProduct] = useState([])
     const dispatch = useDispatch()
     const [loading, setLoading] = useState(true)
+    const [cariProduk, setCariProduk] = useState({ namaProduk: null, kategori: null })
 
-    // const { dataProduct } = useSelector((state) => {
-    //     return {
-    //         dataProduct: state.productReducer.listProduct
-    //     }
-    // })
+    const { dataKategori, dataProductAdmin } = useSelector((state) => {
+        return {
+            dataKategori: state.kategoriReducer.listKategori,
+            dataProductAdmin: state.productReducer.listProductWarehouse
+        }
+    })
     useEffect(() => {
         getData()
-    }, [modalOpen, modalEditOpen])
+    }, [])
     const getData = async () => {
         try {
-            let token = localStorage.getItem('data')
-            let res = await axios.get(`${API_URL}/products/admin`, {
-                headers: {
-                    'Authorization': `Bearer ${token}`
-                }
-            })
-            if (res.data.success) {
-                setDataProduct(res.data.dataProductWarehouse)
-                setLoading(!loading)
-            }
+            setLoading(true)
+            setInterval(() => {
+                setLoading(false)
+            }, 500);
         } catch (error) {
             console.log(error)
         }
     }
     const printProduct = () => {
-        if (dataProduct.length > 0) {
-            return dataProduct.slice(page > 1 ? (page - 1) * limitData : page - 1, page * limitData).map((item, index) => {
+        if (dataProductAdmin.length > 0) {
+            return dataProductAdmin.slice(page > 1 ? (page - 1) * limitData : page - 1, page * limitData).map((item, index) => {
                 return (
                     <>
                         <Tr>
-                            <Th textAlign='center'>{page > 1 ? (page - 1) * limitData + index + 1 : index + 1}</Th>
-                            <Th textAlign='center'>{item.nama}</Th>
-                            <Th textAlign='center'>{item.kategori}</Th>
-                            <Th textAlign='center'>Rp.{item.harga.toLocaleString()}</Th>
-                            <Th textAlign='center'>
+                            <Td textAlign='center' fontWeight='semibold'>{page > 1 ? (page - 1) * limitData + index + 1 : index + 1}</Td>
+                            <Td textAlign='center' fontWeight='semibold'>{item.nama}</Td>
+                            <Td textAlign='center' fontWeight='semibold'>{item.kategori}</Td>
+                            <Td textAlign='center' fontWeight='semibold'>Rp.{item.harga.toLocaleString()}</Td>
+                            <Td textAlign='center'>
                                 <Button colorScheme='yellow' mx='3' size='xs' onClick={() => handleBtEdit(item, true)}><Icon as={AiOutlineEdit} boxSize='20px' color='white' /></Button>
                                 <Button colorScheme='red' size='xs' onClick={() => btDelete(item.idproduct)}><Icon as={AiOutlineDelete} boxSize='20px' /></Button>
-                            </Th>
+                            </Td>
                         </Tr>
                     </>
                 )
@@ -67,16 +66,32 @@ const ManagementProduct = (props) => {
         } else {
             return (
                 <>
-                    <Box my='10vh' display='flex' justifyContent='center'>
-                        <Heading as='h3' size='lg'>Produk tidak ada</Heading>
-                    </Box>
+                    <Tr>
+                        <Td colSpan='5' textAlign='center' fontWeight='semibold'>Produk tidak ada</Td>
+                    </Tr>
                 </>
             )
         }
     }
     const btDelete = async (idproduct) => {
         try {
-            let res = dispatch(deleteProductAction(idproduct))
+            Swal.fire({
+                title: 'Hapus produk ini ?',
+                text: 'Anda tidak dapat mengembalikannya kembali!',
+                showDenyButton: true,
+                confirmButtonColor: 'green',
+                confirmButtonText: 'Konfirmasi',
+                denyButtonText: 'Cancel',
+            }).then((res) => {
+                if (res.isConfirmed) {
+                     dispatch(deleteProductAction(idproduct))
+                    Swal.fire(
+                        'Berhasil!',
+                        'Produk terhapus',
+                        'success'
+                    )
+                }
+            })
         } catch (error) {
             console.log(error)
         }
@@ -89,10 +104,52 @@ const ManagementProduct = (props) => {
         setLimitData(event.target.value)
         setPage(1)
     }
-
+    const btPagination = (event) => {
+        setPage(event)
+        window.scrollTo(0, 0)
+    }
+    const printKategori = () => {
+        return dataKategori.map((item, index) => {
+            return (
+                <>
+                    <option value={item.kategori}>{item.kategori}</option>
+                </>
+            )
+        })
+    }
+    const btFilter = async () => {
+        let data = {
+            ...cariProduk
+        }
+        try {
+            dispatch(getProductWarehouseAction(data))
+        } catch (error) {
+            console.log(error)
+        }
+    }
+    const btReset = async () => {
+        try {
+            dispatch(getProductWarehouseAction())
+            setCariProduk({ namaProduk: '', kategori: '' })
+        } catch (error) {
+            console.log(error)
+        }
+    }
+    const btSorting = async (value) => {
+        let temp = value.split('-')
+        let token = localStorage.getItem('data')
+        let res = await axios.get(`${API_URL}/products/admin?sort=${temp[0]}&order=${temp[1]}`, {
+            headers: {
+                'Authorization': `Bearer ${token}`
+            }
+        })
+        if (res.data.success) {
+            setDataProduct(res.data.dataProductWarehouse)
+        }
+    }
     return (
         <>
-            {/* {console.log('isi detail', detailProduct)} */}
+            {console.log('isi detail', dataProductAdmin)}
             {
                 loading === true ?
                     <LoadingPage />
@@ -107,19 +164,20 @@ const ManagementProduct = (props) => {
                                 <Box w='68vw' my='4vh' p='6' borderRadius='15px' border={'2px solid #F3F4F5'}  >
                                     <Box display='flex' justifyContent='space-between'>
                                         <Box display='flex'>
-                                            <Input placeholder='cari Produk' w='15vw' />
-                                            <Select placeholder='Kategori' w='10vw' mx='5px'>
-                                                <option value='Ruang Keluarga'>Ruang Keluarga</option>
+                                            <Input placeholder='cari Produk' w='15vw' onChange={(event) => setCariProduk({ ...cariProduk, namaProduk: event.target.value })} value={cariProduk.namaProduk} />
+                                            <Select placeholder='Kategori' w='10vw' mx='5px' onChange={(event) => setCariProduk({ ...cariProduk, kategori: event.target.value })} value={cariProduk.kategori}>
+                                                {printKategori()}
                                             </Select>
-                                            <Button colorScheme='telegram'>Filter</Button>
+                                            <Button colorScheme='telegram' onClick={btFilter} mr='5px'>Filter</Button>
+                                            <Button colorScheme='yellow' onClick={btReset}>Reset</Button>
                                         </Box>
                                         <Box>
-                                            <Select placeholder='Sorting berdasarkan'>
+                                            <Select placeholder='Sorting berdasarkan' onChange={(event) => btSorting(event.target.value)}>
                                                 <option value="harga-asc">Harga Asc</option>
                                                 <option value="harga-desc">Harga Desc</option>
                                                 <option value="nama-asc">A - Z</option>
                                                 <option value="nama-desc">Z - A</option>
-                                                <option value="idproduct-asc">Reset</option>
+                                                <option value="added_date-desc">Reset</option>
                                             </Select>
                                         </Box>
                                     </Box>
@@ -156,13 +214,15 @@ const ManagementProduct = (props) => {
                                                     <option value="20">20</option>
                                                     <option value="25">25</option>
                                                 </Select>
-                                                <Pagination total={Math.ceil(dataProduct.length / limitData)} boundaries={2} siblings={2} page={page} onChange={(event) => setPage(event)} />
+                                                <Pagination total={Math.ceil(dataProductAdmin.length / limitData)} boundaries={2} siblings={2} page={page} onChange={(event) => btPagination(event)} />
                                             </InputGroup>
                                         </Center>
                                     </Box>
                                 </Box>
                             </Box>
                         </Box>
+                        <GoOnTop />
+                        <BtnOnTop />
                     </Box>
             }
         </>
